Move pg_bench macro implementation into its own module
Extract the new #[pg_bench] parser and expansion logic from
pgrx-macros/src/lib.rs into pgrx-macros/src/pg_bench.rs, leaving
lib.rs with just the proc-macro entrypoint and delegation.

This addresses the PR review feedback that the amount of new benchmark
macro code warranted its own file. The refactor is intentionally
structural: it preserves the macro surface and generated code while
making the new benchmark-specific parsing and wrapper generation logic
much easier to read and maintain in isolation.

Verification:
- cargo check -p pgrx-macros
- cargo check --manifest-path pgrx-examples/benching/Cargo.toml --no-default-features --features 'pg16 pg_bench'
